requirements.txt
fastapi
uvicorn
stripe
sqlmodel
python-dotenv
email-validator
playwright